Removed forceDoctype, changed default doctype
With the forceDoctype option removed, the compiler now gives presedence
to the doctype set in the view. The default doctype is used if no
doctype is set in the view, or the empty doctype declaration is used. In
the case that no doctype is specified, the default will be added before
the html tag if it is the first tag in the view.