Pass mixin blocks as `this`
By using `Function.call`, we can pass mixin blocks without affecting arguments, which otherwise leads to weird bugs if optional arguments are being used in the mixin.

This also makes the `block` variable true/null instead of returning the function itself, which is safer for users.  The function is instead available as `this.block`.