Stop concurrent jade renders from clobbering each other.
Commit ef156a85bc8 has the side-effect of sharing the runtime
variable between concurrently rendering templates. 

This leads to the exception 'Cannot read property 'filename' of undefined" when one jade render is executed from within another. Before this commit, this used to work. I assume this happens because the debug info clashes.

The easiest fix I could come up with is to give each template it's own 'runtime' object to store debug variables in.