Adds a new `context` keyword.
It is similar to `include`, but instead of including the passed file's contents into the buffered output, the file's context are inserted directly into the compiled template code. This allows you to insert arbitrary javascript into the context of the template, such as JSON data, a bunch of helper javascript functions, etc.

This is particularly useful for things like static site generators that do not have some server middleware to pass in `locals` to the jade template.