Allow spread operator in attributes - N.B. for jsx transform only (#3152)
Currently we only support spread attributes in the transform for producing jsx. Although this adds support for them at the lexer level, they will just be blindly included in the output html when using the standard pug instance. In the future we may introduce proper spread properties to pug as well.