Removed irrelevant info
When I first saw the readme page I thought it was showing me the wrong version of Pug, I really thought I was looking at the first beta v2 from 2016 accidentally but then I realized that, no, I was looking at the current readme page but it just contains outdated info.

To improve the experience for new users I think info about Pug being called Jade seven years ago and info about the beta v2 should be removed from the readme or at least not be the very first thing on the readme. That info is irrelevant to 99% of users. Most new users will use the latest version of Pug, v3.0.2, and they don't need to know or care that Pug was formerly known as Jade in 2016.